+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+### Fixed
+- Keepalive and check TTL failure events now fire continuously until resolved.
+
 ### [5.0.0] - 2018-11-30
 
 ### Added